requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
matplotlib==3.8.2
pandas==2.1.4
tensorflow==2.15.0

